refactor: Make package structure looks better
